fix(codecs)!: use same framing methods for protobuf enc/dec
Having them both use different framing means thats a protobuf source is
incompatible with sink with the 'minimal' config. This is unexpected and
hard to troubleshoot, as you have to look at the http payload to see it.

Length based seems to be the option the protobuf docs recommend:

> Streaming Multiple Messages
> ...
> The easiest way to solve this problem is to write the size of each message before you write the message itself. When you read the messages back in, you read the size, then read the bytes into a separate buffer, then parse from that buffer.
>
> - https://protobuf.dev/programming-guides/techniques/

As this changes framing behaviour of the sink its marked as breaking
change.

Signed-off-by: bas smit <bas@baslab.org>